Refactor: 수정 기존 LoginActivity -> AuthActivity
