+++ app/src/main/res/layout/activity_auth.xml
@@ -4,4 +4,4 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity" />
+    tools:context=".ui.auth.AuthActivity" />
